+++ NEWS.md
@@ -1,6 +1,12 @@
-## qtl2plot 0.5-18 (2017-11-17)
+## qtl2plot 0.5-18 (2017-11-21)
 
 ### New features
+
+- `plot_coef()` now takes an argument `legend` to control addition of
+  a legend. If NULL, no legend is included. Otherwise it indicates the
+  placement of the legend (for example `"topright"` or `"bottomleft"`.
+  There are hidden arguments `legend_lab` for labels and
+  `legend_ncol` for the number of columns.
 
 - New (hidden) graphics paramater for `plot_scan1()`, `altcol`, for
   having colors alternate between chromosomes, for the case of a